LPS-120270 Track and configure portal AuthVerifierPipeline
