+++ modules/apps/portal-security/portal-security-auth-verifier/src/main/java/com/liferay/portal/security/auth/verifier/internal/basic/auth/header/BasicAuthHeaderAuthVerifierPipelineConfigurator.java
@@ -12,10 +12,11 @@
  * details.
  */
 
-package com.liferay.portal.security.auth.verifier.internal.tunnel;
+package com.liferay.portal.security.auth.verifier.internal.basic.auth.header;
 
 import com.liferay.portal.kernel.security.auth.verifier.AuthVerifier;
-import com.liferay.portal.security.auth.verifier.internal.BaseAuthVerifierPublisher;
+import com.liferay.portal.kernel.security.auto.login.AutoLogin;
+import com.liferay.portal.security.auth.verifier.internal.BaseAuthVerifierPipelineConfigurator;
 
 import java.util.Map;
 
@@ -23,30 +24,26 @@ import org.osgi.framework.BundleContext;
 import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.ConfigurationPolicy;
-import org.osgi.service.component.annotations.Deactivate;
-import org.osgi.service.component.annotations.Modified;
+import org.osgi.service.component.annotations.Reference;
 
 /**
  * @author Tomas Polesovsky
  */
 @Component(
-	configurationPid = "com.liferay.portal.security.auth.verifier.internal.tunnel.configuration.TunnelAuthVerifierConfiguration",
+	configurationPid = "com.liferay.portal.security.auth.verifier.internal.basic.auth.header.configuration.BasicAuthHeaderAuthVerifierConfiguration",
 	configurationPolicy = ConfigurationPolicy.OPTIONAL, service = {}
 )
-public class TunnelAuthVerifierPublisher extends BaseAuthVerifierPublisher {
+public class BasicAuthHeaderAuthVerifierPipelineConfigurator
+	extends BaseAuthVerifierPipelineConfigurator {
 
 	@Activate
 	@Override
 	protected void activate(
 		BundleContext bundleContext, Map<String, Object> properties) {
 
-		super.activate(bundleContext, properties);
-	}
+		_authVerifier = new BasicAuthHeaderAuthVerifier(_autoLogin);
 
-	@Deactivate
-	@Override
-	protected void deactivate() {
-		super.deactivate();
+		super.activate(bundleContext, properties);
 	}
 
 	@Override
@@ -54,23 +51,18 @@ public class TunnelAuthVerifierPublisher extends BaseAuthVerifierPublisher {
 		return _authVerifier;
 	}
 
-	@Modified
-	@Override
-	protected void modified(
-		BundleContext bundleContext, Map<String, Object> properties) {
-
-		super.modified(bundleContext, properties);
-	}
-
 	@Override
 	protected String translateKey(String authVerifierPropertyName, String key) {
-		if (key.equals("serviceAccessPolicyName")) {
-			key = "service.access.policy.name";
+		if (key.equals("forceBasicAuth")) {
+			key = "basic_auth";
 		}
 
 		return super.translateKey(authVerifierPropertyName, key);
 	}
 
-	private final AuthVerifier _authVerifier = new TunnelAuthVerifier();
+	private AuthVerifier _authVerifier;
+
+	@Reference(target = "(&(private.auto.login=true)(type=basic.auth.header))")
+	private AutoLogin _autoLogin;
 
 }
