+++ modules/apps/portal-security/portal-security-auth-verifier/src/main/java/com/liferay/portal/security/auth/verifier/internal/digest/authentication/DigestAuthenticationAuthVerifierPipelineConfigurator.java
@@ -15,16 +15,10 @@
 package com.liferay.portal.security.auth.verifier.internal.digest.authentication;
 
 import com.liferay.portal.kernel.security.auth.verifier.AuthVerifier;
-import com.liferay.portal.security.auth.verifier.internal.BaseAuthVerifierPublisher;
+import com.liferay.portal.security.auth.verifier.internal.BaseAuthVerifierPipelineConfigurator;
 
-import java.util.Map;
-
-import org.osgi.framework.BundleContext;
-import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.ConfigurationPolicy;
-import org.osgi.service.component.annotations.Deactivate;
-import org.osgi.service.component.annotations.Modified;
 
 /**
  * @author Tomas Polesovsky
@@ -33,34 +27,12 @@ import org.osgi.service.component.annotations.Modified;
 	configurationPid = "com.liferay.portal.security.auth.verifier.internal.digest.authentication.configuration.DigestAuthenticationAuthVerifierConfiguration",
 	configurationPolicy = ConfigurationPolicy.OPTIONAL, service = {}
 )
-public class DigestAuthenticationAuthVerifierPublisher
-	extends BaseAuthVerifierPublisher {
-
-	@Activate
-	@Override
-	protected void activate(
-		BundleContext bundleContext, Map<String, Object> properties) {
-
-		super.activate(bundleContext, properties);
-	}
-
-	@Deactivate
-	@Override
-	protected void deactivate() {
-		super.deactivate();
-	}
+public class DigestAuthenticationAuthVerifierPipelineConfigurator
+	extends BaseAuthVerifierPipelineConfigurator {
 
 	@Override
 	protected AuthVerifier getAuthVerifierInstance() {
 		return _authVerifier;
-	}
-
-	@Modified
-	@Override
-	protected void modified(
-		BundleContext bundleContext, Map<String, Object> properties) {
-
-		super.modified(bundleContext, properties);
 	}
 
 	@Override
